+++ vm/comp/instructionBlocks/cpu_xorInst.c
@@ -151,9 +151,9 @@ case 0x35:{
 	cpu_print_inst("XOR EAX imm32\n");
 	#endif
 	/*XOR EAX, imm32*/
-	uint8_t op1 = cpu_read_doubleword_from_reg(&(cpu_state->eax));
-	uint8_t op2 = cpu_consume_doubleword_from_mem(cpu_state);
-	uint8_t result = op1 ^ op2;
+	uint32_t op1 = cpu_read_doubleword_from_reg(&(cpu_state->eax));
+	uint32_t op2 = cpu_consume_doubleword_from_mem(cpu_state);
+	uint32_t result = op1 ^ op2;
 	cpu_write_doubleword_in_reg(&(cpu_state->eax), result);
 
 	cpu_clear_flag(cpu_state, OVERFLOW_FLAG);
